adapted format and created sql script to create tables

+++ Lab2TBD/Backend/src/main/java/com/example/Lab1TBD/persistence/entities/ProductEntity.java
@@ -8,11 +8,11 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductEntity {
-    private Long product_id;       // ID UNICO
-    private String product_name;   // nombre del producto
-    private String description;    // descripcion del producto
-    private Float price;           // precio del producto
-    private Integer stock;         // cantidad disponible del producto
-    private String product_status; // estado del producto (disponible,agotado)
-    private Integer category_id;   // identificador foraneo de categoria FK
+    private Long product_id;       // Unique ID
+    private String product_name;   // Product name
+    private String description;    // Product description
+    private Float price;           // Product price
+    private Integer stock;         // Product stock
+    private String product_status; // Product status (available, unavailable)
+    private Long category_id;   // Category ID (FK)
 }
